[refactor] testdata package 생성
- 왜 변경(혹은 추가)이 있었나요?
	- mvc 테스트에서 테스트를 위한 클래스들이 필요했음
	- 그런데 테스트코드와 함께 있으니 복잡함
		- 테스트코드 vs 테스트를 위한 코드
- 무엇이 변경되었나요?
	- testdata package 생성
		- 테스트를 위한 코드를 모아놓는 역할
