+++ nextstep-mvc/src/test/java/nextstep/db/DataBase.java
@@ -1,7 +1,7 @@
 package nextstep.db;
 
 import com.google.common.collect.Maps;
-import nextstep.mvc.tobe.User;
+import nextstep.mvc.tobe.testdata.User;
 
 import java.util.Collection;
 import java.util.Map;
