+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/AnnotationHandlerMappingTest.java
@@ -3,6 +3,7 @@ package nextstep.mvc.tobe;
 import nextstep.db.DataBase;
 import nextstep.mvc.handlermapping.AnnotationHandlerMapping;
 import nextstep.mvc.handlermapping.HandlerExecution;
+import nextstep.mvc.tobe.testdata.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
