+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/HandlerMethodArgumentResolverTest.java
@@ -1,5 +1,6 @@
 package nextstep.mvc.tobe;
 
+import nextstep.mvc.tobe.testdata.TestUserController;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
