+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/testdata/Car.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.testdata;
 
 import java.util.Objects;
