+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/testdata/MyController.java
@@ -1,6 +1,7 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.testdata;
 
 import nextstep.db.DataBase;
+import nextstep.mvc.tobe.ModelAndView;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
