+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/testdata/TestUser.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.testdata;
 
 public class TestUser {
     private String userId;
